% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime_damage}
\alias{crime_damage}
\title{A synthetic data of crime damage.}
\format{
A data frame with 250 rows and 6 variables:
\describe{
\item{collective_efficacy}{}
\item{unemployment}{}
\item{median_age}{}
\item{white_british}{}
\item{damage_crime}{}
\item{log_damage_crime}{}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
crime_damage
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
