% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcme_ind_univar.R
\name{rcme_ind_univar}
\alias{rcme_ind_univar}
\title{Function simulating measurement error on an independent variable}
\usage{
rcme_ind_univar(data, outcome, S, key_predictor, R_sd, D)
}
\arguments{
\item{D}{}
}
\description{
Function simulating measurement error on an independent variable
}
