% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcme_out.R
\name{rcme_out}
\alias{rcme_out}
\title{Function that simulates the effect of measurement error on the dependent variable.}
\usage{
rcme_out(formula, data, focal_variable, R = 1, D = 0, log_var = FALSE)
}
\arguments{
\item{formula}{Regression formula of interest}

\item{data}{Data that will be used}

\item{focal_variable}{The key independent variable of interest in the model. Only one variable can be used.}

\item{R}{The magnitude of the systematic error component, included as a scalar contained within the (0,1). Users can input multiple values using c().}

\item{D}{The magnitude of the association between the measurement error and the key variable of interest measured as relative risk. Users can input multiple values using c().}

\item{log_var}{Do you want the function to log the outcome? If log_var = TRUE users must remember that all sensitivity results will be presented on the log scale.}
}
\description{
Function that simulates the effect of measurement error on the dependent variable.
}
\examples{
rcme_out(
  formula = "damage_crime ~ collective_efficacy + unemployment +
  white_british + median_age",
  data = crime_damage,
  focal_variable = "collective_efficacy",
  R = 0.29,
  D = 0.9,
  log_var = TRUE)
}
