% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcme_out_univar.R
\name{rcme_out_univar}
\alias{rcme_out_univar}
\title{Function simulating measurement error on an outcome variable}
\usage{
rcme_out_univar(data, outcome, S, key_predictor, R_sd, D)
}
\arguments{
\item{D}{}
}
\description{
Function simulating measurement error on an outcome variable
}
