% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcme_sim_plot.R
\name{rcme_sim_plot}
\alias{rcme_sim_plot}
\title{Title}
\usage{
rcme_sim_plot(rcme_sim_range_object, ci = TRUE, naive = TRUE, rr = FALSE)
}
\arguments{
\item{rcme_sim_range_object}{Object saved by \code{rcme_ind()} or \code{rcme_out()} with multiple values.}

\item{ci}{include confidence interval in the graph}

\item{naive}{include naive estimate in the graph}

\item{rr}{show error on relative risk scale}
}
\description{
Title
}
\examples{
ex <- rcme_out(
  formula = "damage_crime ~ collective_efficacy + unemployment +
  white_british + median_age",
  data = crime_damage,
  focal_variable = "collective_efficacy",
  R = c(0.29, 0.85, 0.9),
  D = c(0.9, 1, 1.1),
  log_var = TRUE)

rcme_sim_plot(ex)
}
