% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime_disorder}
\alias{crime_disorder}
\title{A synthetic data of crime disorder.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{violent_crime}{}
\item{white_british}{}
\item{unemployment}{}
\item{median_age}{}
\item{disorder}{}
\item{log_violent_crime}{}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
crime_disorder
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
