% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcme_ind.R
\name{rcme_ind}
\alias{rcme_ind}
\title{Function that simulates the effect of measurement error on an independent variable}
\usage{
rcme_ind(formula, data, key_predictor, S = 1, R_sd = 0, D = 0, log_var = FALSE)
}
\arguments{
\item{formula}{Regression formula of interest}

\item{data}{Data that will be used}

\item{key_predictor}{The key variable of interest in the model}

\item{S}{An estimate of counting rate}

\item{R_sd}{Standard deviation}

\item{D}{Correlation between measurement error and variables of interest}

\item{log_var}{Do you want the function to log the outcome?}
}
\description{
Function that simulates the effect of measurement error on an independent variable
}
